HighScoresUI&PlayerInput
Updated the UI so that the high scores list the top 10 highest scores. Currently using mixture of scores attained during the game, and a dataset that was loaded in according to course-work I am currently working on for dummy numbers.
Also added an input for the user to put their own name in before the game begins.

Next step is getting the scores to project onto the Game Over canvas after the player has ended their run.